react-table: define basic sorting and filtering functions for general use in react tables
