+++ app/assets/javascripts/Components/student_table.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import {render} from 'react-dom';
 
 import {CheckboxTable, withSelection} from './markus_with_selection_hoc'
+import {stringFilter} from './table_helpers';
 
 
 class RawStudentTable extends React.Component {
@@ -172,6 +173,7 @@ class RawStudentTable extends React.Component {
             }
           ]}
           filterable
+          defaultFilterMethod={stringFilter}
           loading={loading}
 
           {...this.props.getCheckboxProps()}
