+++ app/assets/javascripts/Components/ta_table.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import {render} from 'react-dom';
 
 import ReactTable from 'react-table';
+import {stringFilter} from './table_helpers';
+import {CheckboxTable} from "./markus_with_selection_hoc";
 
 
 class TATable extends React.Component {
@@ -67,6 +69,7 @@ class TATable extends React.Component {
           }
         ]}
         filterable
+        defaultFilterMethod={stringFilter}
       />
     );
   }
